requirements.txt
absl-py==2.2.2
aniso8601==10.0.1
astunparse==1.6.3
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-PyMongo==3.0.1
Flask-RESTful==0.3.10
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.1
grpcio==1.71.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.9
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
protobuf==5.29.4
Pygments==2.19.1
pytz==2025.2
requests==2.32.3
rich==14.0.0
six==1.17.0
SQLAlchemy==2.0.40
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
PyJWT==2.10.1
pymongo==4.12.1
